@@ -1,5 +1,5 @@
 {
-    "project_version": "2.9.1",
+    "project_version": "2.10.0+dev.1",
     "electron_version": "4",
     "name": "valorant-companion",
     "version": "1.0.0",
@@ -27,6 +27,7 @@
         "cryptr": "^6.2.0",
         "electron-squirrel-startup": "^1.0.0",
         "localforage": "^1.10.0",
+        "pako": "^2.1.0",
         "path": "^0.12.7",
         "ws": "^8.13.0",
         "zip-lib": "^0.7.3"
@@ -35,6 +36,7 @@
         "@babel/core": "^7.20.12",
         "@babel/eslint-parser": "^7.12.16",
         "@types/node-fetch": "^2.6.2",
+        "@types/pako": "^2.0.0",
         "@typescript-eslint/eslint-plugin": "^5.4.0",
         "@typescript-eslint/parser": "^5.4.0",
         "@vue/cli-plugin-babel": "~5.0.0",
@@ -42,7 +44,7 @@
         "@vue/cli-service": "~5.0.0",
         "@vue/eslint-config-typescript": "^9.1.0",
         "core-js": "^3.8.3",
-        "electron": "^23.2.2",
+        "electron": "^24.8.3",
         "electron-builder": "^24.0.0",
         "electron-builder-squirrel-windows": "^24.2.0",
         "eslint": "^7.32.0",
@@ -50,7 +52,7 @@
         "npm-run-all": "^4.1.5",
         "prettier-plugin-sort-imports": "^1.7.0",
         "typescript": "^4.7.4",
-        "vue": "^3.2.47",
+        "vue": "^3.3.4",
         "vue-class-component": "^8.0.0-0"
     },
     "eslintConfig": {
